+++ imeditor/interface/interface.py
@@ -476,7 +476,7 @@ class Interface(Gtk.ApplicationWindow):
         dialog = Gtk.AboutDialog(transient_for=self)
         dialog.set_logo(self.logo)
         dialog.set_program_name(self.program_title)
-        dialog.set_version('0.7')
+        dialog.set_version('0.8')
         dialog.set_website('https://imeditor.github.io')
         dialog.set_authors(['Nathan Seva', 'Hugo Posnic'])
         gtk_version = '{}.{}.{}'.format(Gtk.get_major_version(),